+++ test/mocks/github-api.mock.ts
@@ -34,6 +34,18 @@ export const createBlob = jest.fn()
     .mockImplementation(async () => "This is a commit message")
     .mockName("github-api.getCommitMessage");
 
+export const getPrComments = jest.fn()
+  .mockImplementation(() => ({
+    [Symbol.asyncIterator](): unknown {
+      return {
+        async next(): Promise<unknown> {
+          return { done: true };
+        },
+      };
+    },
+  }))
+  .mockName("github-api.getPrFile");
+
 export const getPrFile = jest.fn()
   .mockImplementation(async (_, path) => await getContents(path))
   .mockName("github-api.getPrFile");