+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Disable the action from a Pull Request comment. ([#106])
 
 ## [0.2.2] - 2020-03-02
 
@@ -63,4 +63,5 @@ Versioning].
 [#99]: https://github.com/ericcornelissen/svgo-action/pull/99
 [#100]: https://github.com/ericcornelissen/svgo-action/pull/100
 [#103]: https://github.com/ericcornelissen/svgo-action/pull/103
+[#106]: https://github.com/ericcornelissen/svgo-action/pull/106
 [#108]: https://github.com/ericcornelissen/svgo-action/pull/108